Update tests (#1848)
