+++ apps/browser-extension/src/utils/formDetector/__tests__/FormFiller.nl.test.ts
@@ -3,10 +3,11 @@ import { describe, it, expect, beforeEach, vi } from 'vitest';
 
 import type { Credential } from '@/utils/dist/core/models/vault';
 
+import { FormDetector } from '../FormDetector';
 import { FormFiller } from '../FormFiller';
 import { FormFields } from '../types/FormFields';
 
-import { setupTestDOM, createMockFormFields, createMockCredential, wasTriggerCalledFor, createDateSelects } from './TestUtils';
+import { setupTestDOM, createMockFormFields, createMockCredential, wasTriggerCalledFor, createDateSelects, createTestDom } from './TestUtils';
 
 const { window } = new JSDOM('<!DOCTYPE html>');
 global.HTMLSelectElement = window.HTMLSelectElement;
@@ -60,6 +61,135 @@ describe('FormFiller Dutch', () => {
       expect(wasTriggerCalledFor(mockTriggerInputEvents, daySelect)).toBe(true);
       expect(wasTriggerCalledFor(mockTriggerInputEvents, monthSelect)).toBe(true);
       expect(wasTriggerCalledFor(mockTriggerInputEvents, yearSelect)).toBe(true);
+    });
+  });
+
+  describe('Dutch login form 1 field filling', () => {
+    const htmlFile = 'nl-login-form1.html';
+    let mockCredential: Credential;
+
+    beforeEach(() => {
+      mockCredential = createMockCredential();
+    });
+
+    it('should fill username field successfully', async () => {
+      const dom = createTestDom(htmlFile);
+      const doc = dom.window.document;
+
+      const usernameInput = doc.getElementById('login_form_user') as HTMLInputElement;
+      const formDetector = new FormDetector(doc, usernameInput);
+      const detectedFields = formDetector.getForm();
+
+      expect(detectedFields).not.toBeNull();
+
+      if (detectedFields) {
+        const triggerMock = vi.fn();
+        const filler = new FormFiller(detectedFields, triggerMock);
+        await filler.fillFields(mockCredential);
+
+        expect(usernameInput.value).toBe('testuser');
+        expect(wasTriggerCalledFor(triggerMock, usernameInput)).toBe(true);
+      }
+    });
+
+    it('should fill password field successfully', async () => {
+      const dom = createTestDom(htmlFile);
+      const doc = dom.window.document;
+
+      const passwordInput = doc.getElementById('login_form_password') as HTMLInputElement;
+      const formDetector = new FormDetector(doc, passwordInput);
+      const detectedFields = formDetector.getForm();
+
+      expect(detectedFields).not.toBeNull();
+
+      if (detectedFields) {
+        const triggerMock = vi.fn();
+        const filler = new FormFiller(detectedFields, triggerMock);
+        await filler.fillFields(mockCredential);
+
+        await new Promise(resolve => setTimeout(resolve, 150));
+
+        expect(passwordInput.value).toBe('testpass');
+        expect(wasTriggerCalledFor(triggerMock, passwordInput)).toBe(true);
+      }
+    });
+
+    it('should fill both username and password fields without filling other fields', async () => {
+      const dom = createTestDom(htmlFile);
+      const doc = dom.window.document;
+
+      const usernameInput = doc.getElementById('login_form_user') as HTMLInputElement;
+      const passwordInput = doc.getElementById('login_form_password') as HTMLInputElement;
+      const sessionNameInput = doc.getElementById('login_form_sessionName') as HTMLInputElement;
+      const hiddenLocation = doc.getElementById('login_form_location') as HTMLInputElement;
+
+      const formDetector = new FormDetector(doc, usernameInput);
+      const detectedFields = formDetector.getForm();
+
+      expect(detectedFields).not.toBeNull();
+
+      if (detectedFields) {
+        const triggerMock = vi.fn();
+        const filler = new FormFiller(detectedFields, triggerMock);
+        await filler.fillFields(mockCredential);
+
+        await new Promise(resolve => setTimeout(resolve, 150));
+
+        expect(usernameInput.value).toBe('testuser');
+        expect(passwordInput.value).toBe('testpass');
+
+        expect(sessionNameInput.value).toBe('');
+        expect(hiddenLocation.value).toBe('https://example.com/');
+
+        expect(wasTriggerCalledFor(triggerMock, usernameInput)).toBe(true);
+        expect(wasTriggerCalledFor(triggerMock, passwordInput)).toBe(true);
+        expect(wasTriggerCalledFor(triggerMock, sessionNameInput)).toBe(false);
+        expect(wasTriggerCalledFor(triggerMock, hiddenLocation)).toBe(false);
+      }
+    });
+
+    it('should handle form with credential containing email', async () => {
+      const dom = createTestDom(htmlFile);
+      const doc = dom.window.document;
+
+      const usernameInput = doc.getElementById('login_form_user') as HTMLInputElement;
+      const formDetector = new FormDetector(doc, usernameInput);
+      const detectedFields = formDetector.getForm();
+
+      expect(detectedFields).not.toBeNull();
+
+      if (detectedFields) {
+        const triggerMock = vi.fn();
+        const filler = new FormFiller(detectedFields, triggerMock);
+
+        await filler.fillFields(mockCredential);
+
+        expect(usernameInput.value).toBe('testuser');
+        expect(usernameInput.value).not.toBe('test@example.com');
+      }
+    });
+
+    it('should only fill the actual username field, not hidden fields that might match patterns', async () => {
+      const dom = createTestDom(htmlFile);
+      const doc = dom.window.document;
+
+      const usernameInput = doc.getElementById('login_form_user') as HTMLInputElement;
+      const formDetector = new FormDetector(doc, usernameInput);
+      const detectedFields = formDetector.getForm();
+
+      expect(detectedFields).not.toBeNull();
+
+      if (detectedFields) {
+        const triggerMock = vi.fn();
+        const filler = new FormFiller(detectedFields, triggerMock);
+        await filler.fillFields(createMockCredential());
+
+        const allInputs = Array.from(doc.querySelectorAll('input[type="text"]')) as HTMLInputElement[];
+        const filledInputs = allInputs.filter(input => input.value !== '');
+
+        expect(filledInputs.length).toBe(1);
+        expect(filledInputs[0]).toBe(usernameInput);
+      }
     });
   });
 });
